detect/template: move test to own file in src/tests/
